cli: Extract show command dispatch
The argument parser registers every command surface and delegates git help,
quick actions, file arguments, and replacement input to CLI helpers.

The show subcommand keeps page validation, batch lookup, file-scope
resolution, replacement preview checks, and command invocation inside
parser construction. That makes one command carry workflow dispatch logic
while nearby parser responsibilities live behind module boundaries.

This commit begins tightening parser ownership by moving parsed show
argument dispatch into cli.show_dispatch and wiring the show parser to
dispatch_show_command.

Subsequent commits will retarget show parser tests and add architecture
coverage for the parser boundary.
